drm: panel: cwu50: Fixup for probe ordering and initialization timing
The new cwu50 panel driver in the 6.12.y branch is fragile in terms of
probe ordering. Specifically, the kernel does _not_ by itself guarantee
that the panel is probed _after_ its supply voltage regulators (`iovcc`
and `vci`), and on some hardware, this constantly happens before the
voltage regulators are initialized and brought online.

However the initialization sequence, which can happen immediately after
probe, is only expected to work _after_ `iovcc` and `vci` are brought
online for `tRPWIRES`. This results in unreliable panel behavior, and in
the case of my hardware, non-functional panel for almost every single boot.

This patch brings back probing for `iovcc` and `vci`, and bubbles up
`EPROBE_DEFER` whenever the voltage regulators are not ready. The kernel
will then defer probing and retry later. Since we only succeed if both
of the regulators are probed successfully, this guarantees that the
panel initialization sequence only happens after the supply voltage is
applied.

This patch makes a few further fixes:

1. The SLPOUT and DSiPON commands in the old panel's init sequence are
   unnecessary -- they are included in the 2 mipi commands we call after
the initialzation sequence.
2. The initial reset cycling is brought back to match the init sequence
   with JD9365D datasheet, even though it may not be required every
time. To maintain compatibility with the new panel, the reset pin (or ID
pin for the new panel) is read once during probe to detect whether we
are using the new or old panel. That is then stored as a bool. The GPIO
pin is then put into output mode to be used for resetting, in case of
the old panel.
3. Even though the panel's regulartors are marked as always-online in the
   device tree, this commit makes it so that the driver will still
attempt to call `regulator_enable()` on the regulators. This ensures
that the driver works without assumption of how the rest of the device
tree is defined.
4. The `iovcc` regulator is now also added to the CM4 device tree. I'm
   not sure what `vci` is supposed to be on CM4, and since I don't have
a device to test, I'm going to leave it like that for now. Having
`iovcc` at least ensures probe ordering to be corrrect on CM4 too.
5. `unprepare` code is brought back so that the panel is actually put to
   sleep before being blanked.

As a side node, I am ~80% sure that there is no reason to have a
separate driver for CM3 either. As long as probe ordering is maintained
like we have after this patch, I can't see why the same driver wouldn't
work for CM3 either. This is probably true for DevTerm as well. However
since I don't have the corresponding hardware to test with, I'm going to
leave that out. If you do have a CM3 at hand, I think it makes sense to
try to see if the same driver works there as well.