arch: arm: dts: Implement headphone detection and switching for CM5 (#4)
The CM5 makes use of `simple-audio-card`, which actually has support for
headphone pin detection and amplifier switching through
`simple-amplifier` out of the box. This commit simply adds that to the
device tree, so that no userspace daemon is necessary for headphone
switching to work on CM5.

Note that because we need to _disable_ an amplifier when headphone is
plugged in, the usage of `simple-amplifier` is sort of reversed. This is
fine though as long as that is the only thing we use it for.

Unfortunately, I do not think there is an easy way to adapt this for CM3
and CM4. However, this change doesn't conflict with the userspace
switching implementation, so there shouldn't be any issue with having
both this and the userspace script. It should also be quite
straightforward to make the userspace script a no-op on CM5.